Respect rate limiting for GitHub query
GitHub allows 30 requests per minute for the search API which delivers results in chunks of 30 items.
This resolves approximately to 1 item every 0.06 seconds.
Added a pause which should ensure we stay below this limit.

Currently, there are not that many tools for Timewarrior, but it is good to be prepared.