Replace Timewarrior tutorial
This replaces the old one-page tutorial with a more structured user journey.
The new tutorial tries to give a short introduction to the main aspects of how to use Timewarrior
and focuses more on beginners and first time users.
Advanced topics like exclusions, holidays, or themes are therefore left out.
Those and others will be covered in separate tutorials later.

Signed-off-by: Thomas Lauf <thomas.lauf@tngtech.com>